- rename mode1_mode2 example - fix values in normalize_x functions
