+++ src/normalize_x.jl
@@ -18,14 +18,13 @@ end
     end
 end
 
-
-_normalize_x_value(::typeof(compute_stress), char_stress, char_strainrate) = char_strainrate
-_normalize_x_value(::typeof(compute_pressure), char_stress, char_strainrate) = char_strainrate
-_normalize_x_value(::typeof(compute_strain_rate), char_stress, char_strainrate) = char_stress
-_normalize_x_value(::typeof(compute_volumetric_strain_rate), char_stress, char_strainrate) = char_stress
+_normalize_x_value(::typeof(compute_stress), char_stress, char_strainrate) = char_stress
+_normalize_x_value(::typeof(compute_pressure), char_stress, char_strainrate) = char_stress
+_normalize_x_value(::typeof(compute_strain_rate), char_stress, char_strainrate) = char_strainrate
+_normalize_x_value(::typeof(compute_volumetric_strain_rate), char_stress, char_strainrate) = char_strainrate
 _normalize_x_value(::typeof(compute_lambda), char_stress, char_strainrate) = char_stress
 _normalize_x_value(::typeof(compute_lambda_parallel), char_stress, char_strainrate) = char_stress
-_normalize_x_value(::typeof(compute_plastic_strain_rate), char_stress, char_strainrate) = char_stress
+_normalize_x_value(::typeof(compute_plastic_strain_rate), char_stress, char_strainrate) = char_strainrate
